refactor: restructure WolframTool to tools/wolfram directory
- Move WolframTool directory to tools/wolfram for better organization
- Update all import paths across the codebase
- Add tools path alias to tsconfig.json and webpack config
- Maintain functionality while improving code structure
