+++ src/commands/cleanCommands.ts
@@ -2,7 +2,7 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
 
-import { ProgressViewProvider } from '../progressView/ProgressViewProvider';
+import { ProgressViewProvider } from '@progressView/ProgressViewProvider';
 
 // Local imports - log
 import * as logger from '@logger/logUtils';
