+++ src/commands/executeCommand.ts
@@ -2,11 +2,11 @@
 import * as vscode from 'vscode';
 
 // Local imports - agent components
-import { AgentConfig } from '../agent/core/AgentConfig';
-import { executeAgent } from '../agent/runtime/executeAgent';
+import { AgentConfig } from '@agent/core/AgentConfig';
+import { executeAgent } from '@agent/runtime/executeAgent';
 
 // Local imports - history
-import { AgentHistoryManager } from '../historyView/AgentHistoryManager';
+import { AgentHistoryManager } from '@historyView/AgentHistoryManager';
 
 // Add the registration function
 export function registerExecuteCommand(context: vscode.ExtensionContext) {
