+++ src/webview/MessageHandler.ts
@@ -40,10 +40,10 @@ import {
 import { sleep } from '@utils/helpers';
 
 // Local imports - agent
-import { ToolConfig } from '../agent/core/ToolConfig';
-import { AgentConfig } from '../agent/core/AgentConfig';
-import { getAgentPath } from '../agent/runtime/executeAgent';
-import { loadAgentSettingAndPrompts } from '../agent/runtime/agentLoad';
+import { ToolConfig } from '@agent/core/ToolConfig';
+import { AgentConfig } from '@agent/core/AgentConfig';
+import { getAgentPath } from '@agent/runtime/executeAgent';
+import { loadAgentSettingAndPrompts } from '@agent/runtime/agentLoad';
 
 const CHANNEL = 'MessageHandler';
